+++ CHANGELOG.md
@@ -3,6 +3,7 @@ yii2-dynamicform change Log
 
 dev-master
 ----------
+- enh: Added support for kartik-v/yii2-widget-timepicker
 - enh: Added support for kartik-v/yii2-widget-colorinput